requirements.txt
Flask>=2.3
requests>=2.31
psycopg2-binary>=2.9
python-dotenv>=1.0